+++ docs/source/history.rst
@@ -13,6 +13,9 @@ V1.7.0
 ADC/DAC:
     * Fixed bug. When GP1, 2 and 3 are all in ADC mode and ADC reference is VDD. If just after reset VRM reference is selected, ADC stops working.
 
+I2C:
+    * Removed deprecated functions *I2C_is_idle* and *I2C_cancel*.
+
 Misc:
     * Solved bug: USB remote Wake-up is not saved with :func:`save_config`.
     * New behavior for :func:`enable_power_management`. Changes are not saved immediately to Flash. Call to :func:`save_config` is needed instead.